+++ src/ewokscore/_serialization/common/utils/traversal.py
